Remove the MaximumRangeBurningAttack + minor improvement to Enneries Alt max range implementation
Norse Fortress has the Norse Burning Pitch Upgrade. This is being sent as a Ranged attack but the Range attack was providing another entry into the Maximum range of the items illogically.

I set this to be same as MaximumRange to have only one entry in the gear.

Also made sure that the Normal MaximumRange of the arrows affects the Alt Range of the Enneries instead of MaximumRange2. MaximumRange2 still is different to MaximumRange for display purposes in Stats.